fix: cosmos modularization
