+++ src/lib/components/Swap/Swap.fixture.tsx
@@ -1,13 +1,7 @@
-import { SupportedChainId } from 'constants/chains'
-import { DEFAULT_LOCALE, SUPPORTED_LOCALES } from 'constants/locales'
 import { useAtom } from 'jotai'
-import { useAtomValue, useUpdateAtom } from 'jotai/utils'
-import { URLS } from 'lib/cosmos/connectors/network'
-import { cosmosProviderAtom } from 'lib/cosmos/state/provider'
-import { cosmosThemeAtom } from 'lib/cosmos/state/theme'
-import { cosmosWidthAtom } from 'lib/cosmos/state/width'
+import { useUpdateAtom } from 'jotai/utils'
 import { useEffect } from 'react'
-import { useSelect, useValue } from 'react-cosmos/fixture'
+import { useValue } from 'react-cosmos/fixture'
 
 import Swap from '.'
 import { colorAtom } from './Output'
@@ -23,16 +17,6 @@ const validateColor = (() => {
 })()
 
 function Fixture() {
-  const width = useAtomValue(cosmosWidthAtom)
-  const theme = useAtomValue(cosmosThemeAtom)
-  const provider = useAtomValue(cosmosProviderAtom)
-  const [jsonRpcEndpoint] = useSelect('jsonRpcEndpoint', {
-    defaultValue: URLS[SupportedChainId.MAINNET][0],
-    options: Object.values(URLS).flat(),
-  })
-
-  const [locale] = useSelect('locale', { defaultValue: DEFAULT_LOCALE, options: ['pseudo', ...SUPPORTED_LOCALES] })
-
   const [input, setInput] = useAtom(inputAtom)
   const [output, setOutput] = useAtom(outputAtom)
   const [swap, setSwap] = useAtom(swapAtom)
@@ -77,7 +61,7 @@ function Fixture() {
     }
   }, [color, setColor])
 
-  return <Swap jsonRpcEndpoint={jsonRpcEndpoint} locale={locale} provider={provider} theme={theme} width={width} />
+  return <Swap />
 }
 
 export default <Fixture />
