+++ src/lib/index.tsx
@@ -1,1 +1,12 @@
-export { default as Swap } from './components/Swap'
+import Swap from './components/Swap'
+import Widget, { WidgetProps } from './components/Widget'
+
+type SwapWidgetProps = Omit<WidgetProps, 'children'>
+
+export function SwapWidget(props: SwapWidgetProps) {
+  return (
+    <Widget {...props}>
+      <Swap />
+    </Widget>
+  )
+}
